+++ api/checkout/checkout.model.js
@@ -21,11 +21,11 @@ const CheckoutSchema = mongoose.Schema(
       trim: true,
       uppercase: true,
     },
-    // userId: {
-    //   type: mongoose.Schema.Types.ObjectId,
-    //   ref: 'User',
-    //   required: true,
-    // },
+    userId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'User',
+      required: true,
+    },
   },
   {
     timestamps: true,
