+++ api/checkout/index.js
@@ -2,8 +2,10 @@ const { Router } = require('express');
 
 const { handlerCheckout } = require('./checkout.controller');
 
+const { validateJwtMw } = require('../../middlewares/tokenValidator');
+
 const router = Router();
 
-router.post('/', handlerCheckout);
+router.post('/', [validateJwtMw], handlerCheckout);
 
 module.exports = router;
